Changing pages to ssr part 6: forbidden [nodeploy]
